moved src/ to geonoderest and introduced new setup script
