+++ geonodectl
@@ -8,16 +8,16 @@ from typing import List, Union
 from argparse import RawTextHelpFormatter
 from pathlib import Path
 
-from src.geonodetypes import GeonodeEnv
-from src.geonodeobject import GeonodeObjectHandler
-from src.datasets import GeonodeDatasetsHandler
-from src.resources import GeonodeResourceHandler, SUPPORTED_METADATA_TYPES, DEFAULT_METADATA_TYPE
-from src.documents import GeonodeDocumentsHandler
-from src.maps import GeonodeMapsHandler
-from src.people import GeonodePeopleHandler
-from src.geoapps import GeonodeGeoappsHandler
-from src.uploads import GeonodeUploadsHandler
-from src.executionrequest import GeonodeExecutionRequestHandler
+from geonoderest.geonodetypes import GeonodeEnv
+from geonoderest.geonodeobject import GeonodeObjectHandler
+from geonoderest.datasets import GeonodeDatasetsHandler
+from geonoderest.resources import GeonodeResourceHandler, SUPPORTED_METADATA_TYPES, DEFAULT_METADATA_TYPE
+from geonoderest.documents import GeonodeDocumentsHandler
+from geonoderest.maps import GeonodeMapsHandler
+from geonoderest.people import GeonodePeopleHandler
+from geonoderest.geoapps import GeonodeGeoappsHandler
+from geonoderest.uploads import GeonodeUploadsHandler
+from geonoderest.executionrequest import GeonodeExecutionRequestHandler
 
 GEONODECTL_URL_ENV_VAR: str = "GEONODECTL_URL"
 GEONODECTL_BASIC_ENV_VAR: str = "GEONODECTL_BASIC"
