+++ geonoderest/executionrequest.py
@@ -1,8 +1,8 @@
 from typing import Dict, List
 
-from src.geonodetypes import GeonodeCmdOutListKey, GeonodeCmdOutObjectKey
-from src.rest import GeonodeRest
-from src.cmdprint import print_list_on_cmd, print_json
+from geonoderest.geonodetypes import GeonodeCmdOutListKey, GeonodeCmdOutObjectKey
+from geonoderest.rest import GeonodeRest
+from geonoderest.cmdprint import print_list_on_cmd, print_json
 
 
 class GeonodeExecutionRequestHandler(GeonodeRest):
