+++ geonoderest/geonodeobject.py
@@ -1,9 +1,9 @@
 from typing import List, Dict
 import json
 
-from src.geonodetypes import GeonodeCmdOutObjectKey, GeonodeCmdOutListKey
-from src.rest import GeonodeRest
-from src.cmdprint import print_list_on_cmd, print_json
+from geonoderest.geonodetypes import GeonodeCmdOutObjectKey, GeonodeCmdOutListKey
+from geonoderest.rest import GeonodeRest
+from geonoderest.cmdprint import print_list_on_cmd, print_json
 
 
 class GeonodeObjectHandler(GeonodeRest):
