+++ geonoderest/rest.py
@@ -2,7 +2,8 @@ from typing import List, Dict, Optional, TypeAlias, Callable, Any
 
 import urllib3
 import requests
-from src.geonodetypes import GeonodeEnv, GeonodeHTTPFile
+import logging
+from geonoderest.geonodetypes import GeonodeEnv, GeonodeHTTPFile
 
 urllib3.disable_warnings()
 
@@ -85,6 +86,7 @@ class GeonodeRest(object):
             )
             r.raise_for_status()
         except requests.exceptions.HTTPError as err:
+            logging.error(r.json())
             raise SystemExit(err)
         return r.json()
 
@@ -105,6 +107,7 @@ class GeonodeRest(object):
             r = requests.get(url, headers=self.header, verify=self.verify)
             r.raise_for_status()
         except requests.exceptions.HTTPError as err:
+            logging.error(r.json())
             raise SystemExit(err)
         return r.json()
 
@@ -127,6 +130,7 @@ class GeonodeRest(object):
             r = requests.get(url, headers=self.header, json=params, verify=self.verify)
             r.raise_for_status()
         except requests.exceptions.HTTPError as err:
+            logging.error(r.json())
             raise SystemExit(err)
         return r.json()
 
@@ -151,6 +155,7 @@ class GeonodeRest(object):
             )
             r.raise_for_status()
         except requests.exceptions.HTTPError as err:
+            logging.error(r.json())
             raise SystemExit(err)
         return r.json()
 
@@ -174,6 +179,6 @@ class GeonodeRest(object):
             r = requests.delete(url, headers=self.header, verify=self.verify)
             r.raise_for_status()
         except requests.exceptions.HTTPError as err:
+            logging.error(r.json())
             raise SystemExit(err)
-
         return r.json()
